requirements.txt
requests
nse-api-module